fix(h5): api export error
