+++ packages/taro-h5/src/api/canvas/canvasToTempFilePath.ts
@@ -6,7 +6,7 @@ import { MethodHandler } from '../utils/handler'
  * 把当前画布指定区域的内容导出生成指定大小的图片。在 draw() 回调里调用该方法才能保证图片导出成功。
  * @todo 暂未支持尺寸相关功能
  */
-const canvasToTempFilePath: typeof Taro.canvasToTempFilePath = ({ canvasId, fileType, quality, success, fail, complete }, inst) => {
+export const canvasToTempFilePath: typeof Taro.canvasToTempFilePath = ({ canvasId, fileType, quality, success, fail, complete }, inst) => {
   const handle = new MethodHandler({ name: 'canvasToTempFilePath', success, fail, complete })
   const el = findDOM(inst) as HTMLElement
   const canvas = el?.querySelector(`canvas[canvas-id="${canvasId}"]`) as HTMLCanvasElement
@@ -22,5 +22,3 @@ const canvasToTempFilePath: typeof Taro.canvasToTempFilePath = ({ canvasId, file
     })
   }
 }
-
-export default canvasToTempFilePath
