+++ packages/taro-h5/src/api/canvas/index.ts
@@ -2,9 +2,9 @@ import { temporarilyNotSupport } from '../utils'
 
 // 画布
 export const createOffscreenCanvas = temporarilyNotSupport('createOffscreenCanvas')
-export { default as createCanvasContext } from './createCanvasContext'
-export { default as canvasToTempFilePath } from './canvasToTempFilePath'
-export { default as canvasPutImageData } from './canvasPutImageData'
-export { default as canvasGetImageData } from './canvasGetImageData'
+export * from './createCanvasContext'
+export * from './canvasToTempFilePath'
+export * from './canvasPutImageData'
+export * from './canvasGetImageData'
 
 export const drawCanvas = temporarilyNotSupport('drawCanvas')
