+++ packages/taro-h5/src/api/media/image/index.ts
@@ -4,10 +4,10 @@ import { temporarilyNotSupport } from '../../utils'
 export const saveImageToPhotosAlbum = temporarilyNotSupport('saveImageToPhotosAlbum')
 export const previewMedia = temporarilyNotSupport('previewMedia')
 
-export { previewImage } from './previewImage'
-export { default as getImageInfo } from './getImageInfo'
+export * from './previewImage'
+export * from './getImageInfo'
 
 export const compressImage = temporarilyNotSupport('compressImage')
 export const chooseMessageFile = temporarilyNotSupport('chooseMessageFile')
 
-export { default as chooseImage } from './chooseImage'
+export * from './chooseImage'
