+++ packages/taro-h5/src/api/network/download.ts
@@ -113,7 +113,7 @@ const createDownloadTask = ({ url, header, success, error }): Taro.DownloadTask 
  * 下载文件资源到本地。客户端直接发起一个 HTTPS GET 请求，返回文件的本地临时路径。使用前请注意阅读相关说明。
  * 注意：请在服务端响应的 header 中指定合理的 Content-Type 字段，以保证客户端正确处理文件类型。
  */
-const downloadFile: typeof Taro.downloadFile = ({ url, header, success, fail, complete }) => {
+export const downloadFile: typeof Taro.downloadFile = ({ url, header, success, fail, complete }) => {
   let task: Taro.DownloadTask
   const result: ReturnType<typeof Taro.downloadFile> = new Promise((resolve, reject) => {
     task = createDownloadTask({
@@ -138,5 +138,3 @@ const downloadFile: typeof Taro.downloadFile = ({ url, header, success, fail, co
 
   return result
 }
-
-export default downloadFile
