+++ packages/taro-h5/src/api/network/upload.ts
@@ -142,7 +142,7 @@ const createUploadTask = ({ url, filePath, formData, name, header, timeout, file
 /**
  * 将本地资源上传到服务器。客户端发起一个 HTTPS POST 请求，其中 content-type 为 multipart/form-data。使用前请注意阅读相关说明。
  */
-const uploadFile: typeof Taro.uploadFile = ({ url, filePath, name, header, formData, timeout, fileName, success, fail, complete }) => {
+export const uploadFile: typeof Taro.uploadFile = ({ url, filePath, name, header, formData, timeout, fileName, success, fail, complete }) => {
   let task: Taro.UploadTask
   const result: ReturnType<typeof Taro.uploadFile> = new Promise((resolve, reject) => {
     task = createUploadTask({
@@ -172,5 +172,3 @@ const uploadFile: typeof Taro.uploadFile = ({ url, filePath, name, header, formD
 
   return result
 }
-
-export default uploadFile
